feat: split script.js to several modules
