+++ src/keyState.js
@@ -1,0 +1,10 @@
+module.exports = {
+  shift: false,
+  shiftRight: false,
+  ctrl: false,
+  ctrlRight: false,
+  alt: false,
+  altRight: false,
+  windows: false,
+  caps: false,
+};
